requirements.txt
argparse
pygame
pyserial